dict finished

+++ trunk/dict.h
@@ -1,16 +1,16 @@
 #ifndef _DICT_H
 #define _DICT_H
 
-#include <stddef.h>
+#include <stdint.h>
 
 typedef int dict_bucket_t;
 typedef struct dict_entry dict_entry_t;
 
 typedef struct dict {
-	size_t bucket_size;
+	uint32_t bucket_size;
 	dict_bucket_t *buckets;
-	size_t entry_size;
-	size_t entry_used;
+	uint32_t entry_size;
+	uint32_t entry_used;
 	dict_entry_t *entries;
 	// entry_size - entry_used == number of entries in free_list
 	dict_bucket_t free_list;
